Initial version
Install .zip container like any other Joomla extension. Plugin is configured for a max of 10 subscription groups, but that line can be removed from notifier.xml, which makes the limit 1000 (Joomla default). This plugin is really intended to subscribe mailing lists to receive emails containing either a link or the entire body of articles in categories to which each group is subscribed. It's fairly intuitive how this works, but keep in mind the code doesn't check for you omitting the recipient email address or create an empty group between defined groups. Don't do silly things like that and you'll be fine. ;-)