Class ContentHelperRoute not loaded in back end
The plugin was causing an error "Class ContentHelperRoute not found" only when saving an article in the back end editor. Turns out, that class apparently isn't loaded automatically for back end edits (who knew?). So I had to add a line to check for it and load it to handle this case. Joomla core bug? Maybe. But it's fixed in this plugin.