Standalone scheduleable Joomla task
This task plugin allows you to use the Joomla 4.1+ Task Scheduler to send your content notifications. It uses the Content Notifier Plugin settings so that "Save" and "Task" triggers can be more easily managed and kept mutually exclusive. It therefore depends on the Content Notifier Plugin to be installed to work.